+++ src/index.ts
@@ -30,5 +30,10 @@ export {
 
 export { addEffect, effectDecorator as effect } from './libs/effect';
 
-/** error * */
-export { subscribeError } from './libs/error';
+/** act * */
+export { default as flow } from './libs/flow';
+
+export { Flows } from './libs/flows';
+
+/** experience * */
+export { experience } from './libs/experience';